refactor(resources): move business logic to service layer
