+++ jersey/src/main/java/com/netflix/conductor/server/resources/TaskResource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.conductor.server.resources.v1;
+package com.netflix.conductor.server.resources;
 
 
 import java.util.Comparator;
